Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/xss-protection-parsing-01-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor blocked access to 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl' because the source code of a script was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.


--------
Frame: 'frame'
--------
